Execfile() Calls Removed and Rebuilt w/ Modules

+++ hollowbot.py
@@ -11,22 +11,27 @@
 # - move execfiles to modules
 # - use data->after @ end of getlinks.py to do multiple pages
 # - setup reddit account and integrate api
+# - make conf JSON
 
-import datetime
-import json
-import MySQLdb
-import re
-import time
-import urllib2
-import warnings
+from time import sleep
 
-warnings.filterwarnings('ignore', category = MySQLdb.Warning)
+from conf import _
 
-execfile('init.py')
+# Check to make sure the user actually looked at the conf file
+if 'quit' in _:
+    print("Quit flag found in conf.py (you should probably at least look at the configuration before running, yeah?)")
+    exit(1)
+
+from logging import printLog
+from init import *
+
+import links
+import comments
+import timing
 
 # Crawls URLS from datasources
 for url in _['crawl_urls']:
-    getLinks(url)
+    links.get(url)
 
 #Delete old links
 cur.execute("delete from t3 where created < date_sub(now(), interval %s second)", (_['delete_links_after']))
@@ -36,12 +41,12 @@ db.commit();
 cur.execute("select id, permalink from t3 where last_crawled < date_sub(now(), interval %s second)", (_['recrawl_links_after'],))
 
 for c in cur.fetchall():
-    getComments("http://www.reddit.com" + c[1] + ".json")
+    comments.get("http://www.reddit.com" + c[1] + ".json")
     cur.execute("update t3 set last_crawled = now() where id = %s", (c[0],))
     db.commit()
-    time.sleep(_['sleep'])
+    sleep(_['sleep'])
 
-printStats()
+timing.printStats()
 
 db.close()
 log.close()
